moved CaktNode to treenav_tags module
